+++ src/main/java/com/ericsson/eiffel/remrem/publish/config/PropertiesConfig.java
@@ -8,4 +8,13 @@ public class PropertiesConfig {
 	public static final String USE_PERSISTENCE = "com.ericsson.eiffel.remrem.publish.use.persistence";
 	public static final String CLI_MODE = "com.ericsson.eiffel.remrem.publish.cli.mode";
 	public static final String TEST_MODE = "com.ericsson.eiffel.remrem.publish.cli.test.mode";
+	
+	public static final String INVALID_EVENT_CONTENT = "Invalid event content, client need to fix problem in event before submitting again";
+	public static final String INVALID_MESSAGE = "INVALID_MESSAGE";
+	public static final String SUCCESS = "SUCCESS";
+
+	public static final String EVENT_ID = "eventId";
+	public static final String ID = "id";
+	public static final String META = "meta";
+	public static final String EIFFEL_MESSAGE_VERSIONS = "eiffelMessageVersions";
 }