+++ src/test/java/com/ericsson/eiffel/remrem/publish/cli/CliOptionsUnitTests.java
@@ -52,14 +52,14 @@ public class CliOptionsUnitTests {
     
     @Test
     public void testTlsOption() throws Exception {
-        String[] args = {"-f", "/a/b/c/test.file", "-rk", "test", "-tls", "1"};
+        String[] args = {"-f", "/a/b/c/test.file",  "test", "-tls", "1"};
         CliOptions.parse(args);
         assertTrue(CliOptions.getErrorCodes().isEmpty());
     }
     
     @Test
     public void testTlsOptionFails() throws Exception {
-        String[] args = {"-f", "/a/b/c/test.file", "-rk", "test", "-tlsa", "1.2"};
+        String[] args = {"-f", "/a/b/c/test.file",  "test", "-tlsa", "1.2"};
         CliOptions.parse(args);
         int code = CLIExitCodes.CLI_MISSING_OPTION_EXCEPTION;
         assertTrue(CliOptions.getErrorCodes().contains(code));
@@ -67,7 +67,7 @@ public class CliOptionsUnitTests {
 
     @Test
     public void testTlsVer13OptionFails() throws Exception {
-        String[] args = {"-f", "/a/b/c/test.file", "-rk", "test", "-tls", "1.3"};
+        String[] args = {"-f", "/a/b/c/test.file",  "test", "-tls", "1.3"};
         CliOptions.parse(args);
         int code = CLIExitCodes.CLI_MISSING_OPTION_EXCEPTION;
         assertTrue(CliOptions.getErrorCodes().contains(code));
@@ -75,70 +75,71 @@ public class CliOptionsUnitTests {
 
     @Test
     public void testMbOption() throws Exception {
-        String[] args = {"-f", "/a/b/c/test.file", "-rk", "test", "-mb", "MbInstance1"};
+        String[] args = {"-f", "/a/b/c/test.file",  "test", "-mb", "MbInstance1"};
         CliOptions.parse(args);
+        System.out.println(CliOptions.getErrorCodes());
         assertTrue(CliOptions.getErrorCodes().isEmpty());
     }
     
     @Test
     public void testMbOptionFails() throws Exception {
-        String[] args = {"-f", "/a/b/c/test.file", "-rk", "test", "-mbe", "MbInstance1"};
+        String[] args = {"-f", "/a/b/c/test.file",  "test", "-mbe", "MbInstance1"};
         CliOptions.parse(args);
         int code = CLIExitCodes.CLI_MISSING_OPTION_EXCEPTION;
         assertTrue(CliOptions.getErrorCodes().contains(code));
     }
     
     public void testEnOption() throws Exception {
-        String[] args = {"-f", "/a/b/c/test.file", "-rk", "test", "-en", "exchange_name1"};
+        String[] args = {"-f", "/a/b/c/test.file",  "test", "-en", "exchange_name1"};
         CliOptions.parse(args);
         assertTrue(CliOptions.getErrorCodes().isEmpty());
     }
     
     @Test
     public void testEnOptionFails() throws Exception {
-        String[] args = {"-f", "/a/b/c/test.file", "-rk", "test", "-enf", "exchange_name1"};
+        String[] args = {"-f", "/a/b/c/test.file",  "test", "-enf", "exchange_name1"};
         CliOptions.parse(args);
         int code = CLIExitCodes.CLI_MISSING_OPTION_EXCEPTION;
         assertTrue(CliOptions.getErrorCodes().contains(code));
     }
     
     public void testNpOption() throws Exception {
-        String[] args = {"-f", "/a/b/c/test.file", "-rk", "test", "-np", "non_persistent"};
+        String[] args = {"-f", "/a/b/c/test.file",  "test", "-np", "non_persistent"};
         CliOptions.parse(args);
         assertTrue(CliOptions.getErrorCodes().isEmpty());
     }
     
     @Test
     public void testNpOptionFails() throws Exception {
-        String[] args = {"-f", "/a/b/c/test.file", "-rk", "test", "-npf", "non_persistent"};
+        String[] args = {"-f", "/a/b/c/test.file",  "test", "-npf", "non_persistent"};
         CliOptions.parse(args);
         int code = CLIExitCodes.CLI_MISSING_OPTION_EXCEPTION;
         assertTrue(CliOptions.getErrorCodes().contains(code));
     }
 
     public void testPortOption() throws Exception {
-        String[] args = {"-f", "/a/b/c/test.file", "-rk", "test", "-port", "portA"};
+        String[] args = {"-f", "/a/b/c/test.file",  "test", "-port", "portA"};
         CliOptions.parse(args);
         assertTrue(CliOptions.getErrorCodes().isEmpty());
     }
 
     @Test
     public void testPortOptionFails() throws Exception {
-        String[] args = {"-f", "/a/b/c/test.file", "-rk", "test", "-portf", "portB"};
+        String[] args = {"-f", "/a/b/c/test.file",  "test", "-portf", "portB"};
         CliOptions.parse(args);
         int code = CLIExitCodes.CLI_MISSING_OPTION_EXCEPTION;
         assertTrue(CliOptions.getErrorCodes().contains(code));
     }
     
     public void testFOption() throws Exception {
-        String[] args = {"-f", "/a/b/c/test.file", "-rk", "test", "-port", "portA"};
+        String[] args = {"-f", "/a/b/c/test.file",  "test", "-port", "portA"};
         CliOptions.parse(args);
         assertTrue(CliOptions.getErrorCodes().isEmpty());
     }
 
     @Test
     public void testFOptionFails() throws Exception {
-        String[] args = {"-fab", "/a/b/c/test.file", "-rk", "test", "-portf", "portB"};
+        String[] args = {"-fab", "/a/b/c/test.file",  "test", "-portf", "portB"};
         CliOptions.parse(args);
         int code = CLIExitCodes.CLI_MISSING_OPTION_EXCEPTION;
         assertTrue(CliOptions.getErrorCodes().contains(code));